Added keyed protection
It is NOT ok to allow just anyone to hit this script with any POST that mimics a github deploy hook because it executes shell commands that may or may not be resource intensive.  Therefore I have added in a small measure to protect the script from running fully.  Set your hook to: http://you.crew.ccs.neu.edu/index.php?key=<insert random key here and in script>